Separate polygon utilities from polygon_filter
